updating boilerplate CSV generation with feedbacks from osd-metrics-exporter integration

+++ boilerplate/openshift/golang-osd-operator/csv-generate/common-generate-operator-bundle.py
@@ -45,35 +45,39 @@ with open('config/templates/csv-template.yaml', 'r') as stream:
 csv['spec']['customresourcedefinitions']['owned'] = []
 
 # Copy all CRD files over to the bundle output dir:
-crd_files = [ f for f in os.listdir('deploy/crds') if f.endswith('_crd.yaml') ]
-for file_name in crd_files:
-    full_path = os.path.join('deploy/crds', file_name)
-    if (os.path.isfile(os.path.join('deploy/crds', file_name))):
-        shutil.copy(full_path, os.path.join(version_dir, file_name))
-    # Load CRD so we can use attributes from it
-    with open("deploy/crds/{}".format(file_name), "r") as stream:
-        crd = yaml.load(stream)
-    # Update CSV template customresourcedefinitions key
-    csv['spec']['customresourcedefinitions']['owned'].append(
-        {
-            "name": crd["metadata"]["name"],
-            "description": crd["spec"]["names"]["kind"],
-            "displayName": crd["spec"]["names"]["kind"],
-            "kind": crd["spec"]["names"]["kind"],
-            "version": crd["spec"]["version"]
-        }
-    )
+if ( os.path.exists('deploy/crds') ) : 
+    crd_files = [ f for f in os.listdir('deploy/crds') if f.endswith('_crd.yaml') ]
+    for file_name in crd_files:
+        full_path = os.path.join('deploy/crds', file_name)
+        if (os.path.isfile(os.path.join('deploy/crds', file_name))):
+            shutil.copy(full_path, os.path.join(version_dir, file_name))
+        # Load CRD so we can use attributes from it
+        with open("deploy/crds/{}".format(file_name), "r") as stream:
+            crd = yaml.load(stream)
+        # Update CSV template customresourcedefinitions key
+        csv['spec']['customresourcedefinitions']['owned'].append(
+            {
+                "name": crd["metadata"]["name"],
+                "description": crd["spec"]["names"]["kind"],
+                "displayName": crd["spec"]["names"]["kind"],
+                "kind": crd["spec"]["names"]["kind"],
+                "version": crd["spec"]["version"]
+            }
+        )
 
 csv['spec']['install']['spec']['clusterPermissions'] = []
 
 # Add operator role to the CSV:
-with open('deploy/role.yaml', 'r') as stream:
-    operator_role = yaml.load(stream)
-    csv['spec']['install']['spec']['clusterPermissions'].append(
-        {
-            'rules': operator_role['rules'],
-            'serviceAccountName': operator_name,
-        })
+if ( os.path.exists('deploy/roles') ) : 
+    role_files = [ f for f in os.listdir('deploy/roles') if f.endswith('.yaml') ]
+    for file_name in role_files:
+        with open("deploy/roles/{}".format(file_name), 'r') as stream:
+            operator_role = yaml.load(stream)
+            csv['spec']['install']['spec']['clusterPermissions'].append(
+                {
+                    'rules': operator_role['rules'],
+                    'serviceAccountName': operator_name,
+                })
 
 # Add our deployment spec for the operator:
 with open('deploy/operator.yaml', 'r') as stream:
